backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "amazon-bedrock-haystack>=6.4.0",
    "datasets>=2.6.1",
    "fastapi[standard]>=0.129.0",
    "haystack-ai>=2.24.1",
    "pip-system-certs>=5.3",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=4.1.0",
]
